requirements.txt
aiogram==3.12.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
aiofiles==24.1.0
